Remove GraphQL query from Docs page
Move the query to the DocsRenderer component to avoid unnecessary rerenders
