+++ react/components/Docs/DocsRenderer.tsx
@@ -1,17 +1,34 @@
 import React, { FunctionComponent } from 'react'
 import remark from 'remark'
 import remark2react from 'remark-react'
+import { graphql, compose } from 'react-apollo'
+import { Spinner } from 'vtex.styleguide'
+import { branch, renderComponent } from 'recompose'
 
-const DocsRenderer: FunctionComponent<{ decodedMarkdown: string }> = ({
-  decodedMarkdown,
-}) => (
+import * as RepoDocs from '../../graphql/getDocs.graphql'
+
+const DocsRenderer: FunctionComponent<any> = ({ docsQuery }) => (
   <article className="pa7 bg-base--inverted br3">
     {
       remark()
         .use(remark2react)
-        .processSync(decodedMarkdown).contents
+        .processSync(atob(docsQuery.getDocs.encodedDocs)).contents
     }
   </article>
 )
 
-export default DocsRenderer
+export default compose(
+  graphql(RepoDocs.default, {
+    name: 'docsQuery',
+    options: () => {
+      const params = new URLSearchParams(location.search)
+      const repositoryName = params.get('repo') || 'store-components'
+      return {
+        variables: {
+          repositoryName,
+        },
+      }
+    },
+  }),
+  branch(({ docsQuery }: any) => docsQuery.loading, renderComponent(Spinner))
+)(DocsRenderer)
